tests/wpt/tests/visual-viewport/viewport-scale-clamped.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title>Viewport Scale Clamped: Not Pinch Zoomable</title>
  <link rel=help href="https://github.com/w3c/csswg-drafts/issues/9004">
  <link rel=help href="https://github.com/servo/servo/issues/40390">
  <link rel="author" title="Shubham Gupta" href="mailto:shubham13297@gmail.com">
  <meta name="assert" content="The page can not be pinch-zoom.">
  <meta name="viewport" content="minimum-scale=1, maximum-scale=1">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-actions.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
  <script src="/visual-viewport/viewport_support.js"></script>
</head>

<body>
  <div style="font-size: 50px;">This page can not be pinch zoomed, it is clamped using meta viewport scale.</div>
  <script> promise_test(async () => {
      const initialScale = window.visualViewport.scale;
      // Perform Pinch Zoom-In
      await pinchZoomIn();
      assert_equals(window.visualViewport.scale, initialScale, "Scale should be same");
    }) </script>
</body>

</html>
